changes in data.sql

+++ src/main/java/forex/copytradingforex/service/impl/PictureServiceImpl.java
@@ -5,6 +5,7 @@ import forex.copytradingforex.model.entity.PositionEntity;
 import forex.copytradingforex.model.entity.UserEntity;
 import forex.copytradingforex.model.service.PictureAddServiceModel;
 import forex.copytradingforex.repository.PictureRepository;
+import forex.copytradingforex.service.CloudinaryService;
 import forex.copytradingforex.service.PictureService;
 import forex.copytradingforex.service.PositionService;
 import forex.copytradingforex.service.UserService;
@@ -18,13 +19,15 @@ public class PictureServiceImpl implements PictureService {
     private final ModelMapper modelMapper;
     private final UserService userService;
     private final PositionService positionService;
+    private final CloudinaryService cloudinaryService;
 
-    public PictureServiceImpl(PictureRepository pictureRepository, ModelMapper modelMapper, UserService userService, PositionService positionService) {
+    public PictureServiceImpl(PictureRepository pictureRepository, ModelMapper modelMapper, UserService userService, PositionService positionService, CloudinaryService cloudinaryService) {
         this.pictureRepository = pictureRepository;
         this.modelMapper = modelMapper;
         this.userService = userService;
 
         this.positionService = positionService;
+        this.cloudinaryService = cloudinaryService;
     }
 
     @Override
@@ -33,9 +36,11 @@ public class PictureServiceImpl implements PictureService {
         UserEntity trader = userService.findUserByUsername(pictureAddServiceModel.getTraderUsername());
 
         PositionEntity position = positionService.getPictureById(pictureAddServiceModel.getPositionId());
-        if(position.getPicture()!=null){
-            pictureRepository.delete(position.getPicture());
-            position.setPicture(null);
+        if (position.getPicture() != null) {
+            if (cloudinaryService.delete(position.getPicture().getPublicId())) {
+                pictureRepository.delete(position.getPicture());
+                position.setPicture(null);
+            }
         }
 
         PictureEntity pictureEntity = new PictureEntity()
